requirements.txt
PyMuPDF
fastapi
sentence-transformers
numpy
spacy
python-docx
uvicorn
Groq
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz#egg=en_core_web_sm
python-multipart
pandas